backend/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
psycopg2==2.9.3
django-cors-headers==3.13.0
gunicorn==20.1.0
python-dotenv==1.0.0
coreapi==2.3.3
PyYAML==6.0.2